requirements.txt
Django==3.1.7
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.2
psycopg2-binary
wheel